% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_excelMerger.R
\name{gse_as_df}
\alias{gse_as_df}
\title{GSE in global environment to df}
\usage{
gse_as_df()
}
\description{
Creates a data frame of important information of all the studies compiled, for further processing/export.
}
