% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_studyLoader.R
\name{process_GEO_data}
\alias{process_GEO_data}
\title{process GEO data}
\usage{
process_GEO_data(GSE_vector)
}
\description{
process GEO data by extracting crucial info and unpacks them into global environment.
}
