% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_excelMerger.R
\name{row_creator}
\alias{row_creator}
\title{Creates "rows"}
\usage{
row_creator(GSE_object)
}
\description{
Returns important info of GEO object as a vector, for compilation into a data frame. Modify to suite your needs.
}
