% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_studyLoader.R
\name{string_to_vector}
\alias{string_to_vector}
\title{String to vector converter}
\usage{
string_to_vector(GSE_string)
}
\description{
Returns a string of GSE accession numbers as a vector, for processing.
}
